update build os.xml function

+++ com.mwos.ebochs/src/com/mwos/ebochs/core/build/ProjectBuilder.java
@@ -2,13 +2,11 @@ package com.mwos.ebochs.core.build;
 
 import java.io.File;
 import java.io.IOException;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
 import javax.xml.parsers.ParserConfigurationException;
 
-import org.eclipse.core.resources.ICommand;
 import org.eclipse.core.resources.IProject;
 import org.eclipse.core.resources.IResourceDelta;
 import org.eclipse.core.resources.IncrementalProjectBuilder;
@@ -18,8 +16,12 @@ import org.xml.sax.SAXException;
 
 import com.mwos.ebochs.core.ExceptionUtil;
 import com.mwos.ebochs.core.FileUtil;
-import com.mwos.ebochs.resource.OSConfig;
-import com.mwos.ebochs.resource.OSConfigFactory;
+import com.mwos.ebochs.resource.config.OSConfigFactory;
+import com.mwos.ebochs.resource.config.entity.CodePart;
+import com.mwos.ebochs.resource.config.entity.CodePart.Code;
+import com.mwos.ebochs.resource.config.entity.Image;
+import com.mwos.ebochs.resource.config.entity.OSConfig;
+import com.mwos.ebochs.resource.config.entity.Resource;
 import com.mwos.ebochs.ui.preference.OSDevPreference;
 import com.mwos.ebochs.ui.view.ConsoleFactory;
 
@@ -50,6 +52,7 @@ public class ProjectBuilder extends IncrementalProjectBuilder {
 	@Override
 	protected void clean(IProgressMonitor monitor) throws CoreException {
 		super.clean(monitor);
+		cleanAll(this.getProject());
 	}
 
 	private void doBuilds(IResourceDelta deltas[]) {
@@ -73,48 +76,79 @@ public class ProjectBuilder extends IncrementalProjectBuilder {
 		}
 	}
 
-	private void doBuildAsm(String file) {
+	private boolean doBuildAsm(String file) {
 		try {
+			clean(file, this.getProject());
 			BuildResult res = Compiler.compileAsm(file, this.getProject());
 			if (res.isSuccess()) {
 				ConsoleFactory.outMsg(file + " ----- 编译成功\r\n" + res.getMsg(), this.getProject());
 			} else {
 				ConsoleFactory.outErrMsg(file + " ----- 编译错误\r\n" + res.getAllMsg(), this.getProject());
 			}
+			return res.isSuccess();
 		} catch (Exception e) {
 			e.printStackTrace();
 			ConsoleFactory.outErrMsg(file + " ----- 编译异常\r\n", this.getProject());
+			return false;
 		}
 	}
 
-	private void doBuildC(String file) {
+	private boolean doBuildC(String file) {
 		try {
+			clean(file, this.getProject());
 			BuildResult res = Compiler.compileC(file, this.getProject());
 			if (res.isSuccess()) {
 				ConsoleFactory.outMsg(file + " ----- 编译成功\r\n" + res.getMsg(), this.getProject());
 			} else {
 				ConsoleFactory.outErrMsg(file + " ----- 编译错误\r\n" + res.getAllMsg(), this.getProject());
 			}
+			return res.isSuccess();
 		} catch (Exception e) {
 			e.printStackTrace();
 			ConsoleFactory.outErrMsg(file + " ----- 编译异常\r\n", this.getProject());
+			return false;
 		}
 	}
 
-	private void doBuildOSXml() {
+	private boolean doBuildFile(String file) {
+		if (file.endsWith(".c")) {
+			return doBuildC(file);
+		} else if (file.endsWith(".asm")) {
+			return doBuildAsm(file);
+		}
+
+		return false;
+	}
+
+	private boolean doBuildOSXml() {
 		try {
-			OSConfig config = OSConfigFactory.getConfig(this.getProject());
-			cleanAll(getProject());
-			HashMap<String, OSConfig.Source> sources = config.getSources();
-			for (String code : sources.keySet()) {
-				if (code.endsWith(".asm"))
-					doBuildAsm(code);
-				else if (code.endsWith(".c"))
-					doBuildC(code);
+			OSConfig config = OSConfigFactory.getBuildConfig(this.getProject());
+			if (!config.needBuild())
+				return false;
+			for (Image image : config.getImages()) {
+				for (Resource resource : image.getResources()) {
+					if (resource.getType().equals(Resource.CODEPART)) {
+						CodePart codePart = config.getCodePart(resource.getName());
+						if (codePart == null) {
+							ConsoleFactory.outErrMsg(resource.getName() + " ----- 构建异常：未找到资源\r\n", this.getProject());
+							return false;
+						}
+						for (Code code : codePart.getCodes()) {
+							if (!doBuildFile(code.getName())) {
+								ConsoleFactory.outErrMsg(resource.getName() + " : " + code.getName() + " ----- 构建异常\r\n",
+										this.getProject());
+								return false;
+							}
+						}
+					}
+				}
 			}
+
+			return true;
 		} catch (ParserConfigurationException | SAXException | IOException e) {
-			ConsoleFactory.outErrMsg("OS.xml ----- 构建错误\r\n" + ExceptionUtil.getMsg(e), this.getProject());
+			ConsoleFactory.outErrMsg(this.getProject().getName() + " ----- 构建失败\r\n" + ExceptionUtil.getMsg(e), this.getProject());
 			e.printStackTrace();
+			return false;
 		}
 	}
 
@@ -128,9 +162,11 @@ public class ProjectBuilder extends IncrementalProjectBuilder {
 	private void clean(String name, IProject project) {
 		name = FileUtil.getFileName(name, false);
 		String path = project.getLocationURI().getPath() + "\\obj";
-		List<File> objs = FileUtil.listFiles(path, new String[] { name + "." }, true);
-		for (File f : objs) {
-			f.delete();
-		}
+		// "^aa.*?bb$" aa开头，bb结尾
+		List<File> objs = FileUtil.listFiles(path, new String[] { "^" + name + "\\..*?", "^_" + name + "\\..*?", name }, true);
+		if (objs != null)
+			for (File f : objs) {
+				f.delete();
+			}
 	}
 }
